requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
mojang-api-3==3.0.4
pefile==2019.4.18
PyInstaller==3.6
python-box==4.2.3
pywin32-ctypes==0.2.0
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
toml==0.10.1
urllib3==1.25.9
